requirements.txt
# Core ADK
google-adk
python-dotenv

# LLM Integration
litellm

# RAG Components
langchain
langchain-community
chromadb
llama-index
llama-index-vector-stores-chroma

# Document Processing
PyMuPDF
unstructured
pypdf

# Embeddings
sentence-transformers
huggingface-hub

# Text Processing
tiktoken
